requirements.txt
torch
transformers
pyyaml
tqdm
numpy
bitsandbytes 
accelerate 
optimum 
auto-gptq